requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
dnspython
python-multipart
certifi
